Updated grade data type from int to float, and accounted for edge case in input (input is not int).

+++ gradebook.h
@@ -9,8 +9,8 @@ private:
 	std::string file_name;
 	int num_elements;
 	std::vector<std::string> names_vector;
-	std::vector<int> grades_vector;
-	std::vector<int> grades_total_vector;
+	std::vector<float> grades_vector;
+	std::vector<float> grades_total_vector;
 	std::vector<std::string> categories_vector;
 	std::vector<std::string> courses_vector;
 public:
